requirements.txt
django>=5.0
djangorestframework
psycopg2-binary
drf-spectacular

gunicorn

pytest
pytest-django
coverage

ruff
black
